requirements.txt
pygame==1.9.3
paho-mqtt==1.3.1